+++ agents/nil_v2.py
@@ -268,7 +268,7 @@ class nil_v2_agent(Agent):
                 self.training_steps += 1
         
             outputs.append(output.detach())
-            total_loss += float(torch.mean(loss))
+            total_loss += torch.mean(loss)
 
             torch.cuda.nvtx.range_pop()
